requirements.txt
bitarray==1.6.3
pycryptodome==3.10.1